+++ rmk/src/split/ble/central.rs
@@ -290,8 +290,8 @@ fn defaul_central_conn_param() -> ConnectParams {
     ConnectParams {
         min_connection_interval: Duration::from_micros(7500),
         max_connection_interval: Duration::from_micros(7500),
-        max_latency: 400, // 3s
-        supervision_timeout: Duration::from_secs(7),
+        max_latency: 30, // 225ms
+        supervision_timeout: Duration::from_secs(5),
         ..Default::default()
     }
 }
@@ -557,6 +557,7 @@ async fn sleep_manager_task<
 
             // Connection parameters are different when central is broadcasting and connected to host
             let conn_params = if CONNECTION_STATE.load(Ordering::Acquire) {
+                // Connected, the connection interval is 20ms
                 ConnectParams {
                     min_connection_interval: Duration::from_millis(20),
                     max_connection_interval: Duration::from_millis(20),
@@ -565,6 +566,7 @@ async fn sleep_manager_task<
                     ..Default::default()
                 }
             } else {
+                // Advertising ,the connection interval can be longer
                 ConnectParams {
                     min_connection_interval: Duration::from_millis(200),
                     max_connection_interval: Duration::from_millis(200),