+++ rmk/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add PMW3610 optical mouse sensor support for nRF and RP2040 with bit-bang SPI.
+
+### Changed
+
+- Make `embedded-hal-async` a required dependency.
+
 ## [0.8.1] - 2025-11-25
 
 ### Changed