+++ rmk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.2] - 2025-12-18
+
 ### Fixed
 
 - Fix compilation error when use `Macro()` in keymap config